Adjustments for building with Android NDK
Almost ready for prime time, so start trying to build targetting actual
devices.  clang was worried about ret being unset in testgpt() and it was
right.  So far the if statement has never been false, but in a
theoretical future it could be.
Android's support library looks rather more limited, looks like "UTF8"
and "UTF16LE" are the versions of the locale strings it wants.
Rip out the use of search.h.  Android's version is rather more limited.
While use of linear search is generally a bad thing, since the slices
often line up the cost should be rather low.