Several small, but crucial bugfixes; testing suggests this is operational
Fixed memory clearing after conversion to UTF16LE, needed to be a
character, otherwise addition isn't in bytes.  Provide an alert if the
GPT writing code returns an error (not sure what to do then!).
replacegpt() also needed to ensure memory was cleared.

The on development host testing seems to suggest this version is mostly
functional.  Testers, this is almost ready for you.