Implement internal file copy for dirtysanta
Running `dd` does work, but I'm forced to cringe at calling an outside
program to do a file copy.  This code also checks all return codes to
ensure nothing goes wrong.

I'd like to see whether it is possible for `dirtysanta` to act in the
role of wrapper and invoke `applypatch` itself.  This would simplify the
process and reduce opportunity for user error

Note, this code hasn't yet been tested so bugs are likely.