Some major rework of dirtysanta.c, move backups to dirtysanta_backups
Removed usage of __LARGEFILE64_SOURCE, since we're building on a 64-bit
target this is pretty well unneeded anyway.  Add messages for failed
backups of boot devices.  Try to restore aboot.img if the installation of
the new image fails.  Add support for backing up many other portions of
on-board flash, backup recovery since replacing it is the next step.