Add commented list of potential backup candidates; Add discard support
Someone may want to archive copies of the other portions of flash, so add
notes on their names to the source file.  These are all commented out,
but if the need arises...

Check for existance of the new aboot.img file before trying the write
operation.  This might be used for testing or archiving.

Add support for discarding the end of a target flash device.  For SATA
this would be the "TRIM" command.  Other media uses different variations
for the command, but take advantage of this to go easier on the flash.