Remove unused space at end of images, ensure Python 3 compatibility
Add definitions of additional fields from the boot image header.  Most of
these fields are unimportant, but they do allow computing the actual
significant size of the payload.  With that information, truncate the
extra off of image files.

Check on Python 3 compatibility.  Seems Python 3 has changed "/" to
default to generating a floating-point number, someone needs to be shot
for this.  Fixed.