Make ramdisk file patching be in-order again
The string matches done on files in the ramdisk image were meant to be
in-order, not searching the whole file for each string.  This allowed the
targetting strings by nearby text, though means the entire file isn't
searched.  Since this also speeds things up, return things to the way
they were.

Appears the file wasn't being formally closed when we were done,
implement that now.