Add alternative to iconv, fix block size from BLKSSZGET
Appears iconv really is the right approach for this situation, but with
Android's iconv support being unstable there is a problem.  c16tomb()
and mbrtoc16() aren't nearly as nice, but they do appear to work.  Given
the broken state of Android's iconv, make use of these.

The block size given by BLKSSZGET appears to always be a 32-bit integer.
Adjust ioctl() call and passing of the value to match.