Add KDZ check before template writes, add copyright/license to help message
With test_kdzfile() now operational, now add the check before printing
the messages about "Write X".  Not a big change, but better simulation of
final product.

Adding a brief mention of copyright to the help message.  While I'm not
expecting this program to ever be anywhere where copyright is an issue,
this makes it prominent so people in the right place have a better chance
of spotting the issue.