Add ability to override model/SIM count on command-line
Since it is possible to blank the "misc" area, allow the model name and
SIM count to be specified on the command-line.

Mark arguments meant for use with stock images not to be installed to be
left out of LineageOS.

Allow piping of boot images.  This isn't expected to be the main mode of
operation, but this is a conventional use pattern.