Rename fix-h990-cmdline to fix-h990-modem
The end goal of this program has been to get the H990 modem operational.
Modifying the boot image command-line was a means to this goal, it wasn't
the actual goal.  Since the newer kernel allows this fixup to be done via
other means, it seems appropriate to do a rename.