Fix bug added during reorganization, add extra check
When the reorganization (81331d23e86a5ed86103efbfd7b9270460611b55) was
done, the reads were modified to allow reading/writing from/to pipes.
During this addition, one of the reads got messed up.  Fix this critical
bug.

Add a sanity check for the block size in the boot image.  This seems
unlikely to occur, but that is famous last words.