Add support for eMMC installation, avoid umount when testing
Add support for eMMC devices.  Mostly this simply involves opening
/dev/block/mmcblk[0-9] instead of /dev/block/sd[a-z].  The write process
was done by opening devices in /dev/block/bootdevice/by-name and doesn't
need changing.

Avoid umount of /system if simulating.  The goal is for testing to be
possible when Android is running.  Android really doesn't like /system
being unmounted.

Update copyrights.  This is now 2018, so the modifications have a newer
date.