Loosen KDZ comparison check, ignore likely modified GPT entries
There is a plan for LineageOS to break a /vendor off /system.  As such
start completely ignoring OP, system and userdata GPT entries.  This is a
precursor to restoring GPTs.

There is also some potential for modding the raw_resources area.  This
has been used as part of a bootloader breaking strategy.  Someone might
also want to do something to the LGE startup logos.