Greatly improve entryStart calculation, correct AltLBA interpretation
The previously used computation may have worked due to luck, but this one
is much better.  I'm pretty hopeful this one is Correct(tm).  Include an
extra sanity check.

Rereading the specification, the AlternateLBA is the LBA of the alternate
GPT header, not the LBA of the /other/ header.  ie it should always point
at the last LBA, even in the alternate header.

Typically the slice list is placed reserving space for 128 entries.  This
is a bit silly on modern hardware where moving a 16KB chunk around is
pretty trivial.  Yet most setups still lay things out with that
limitation.  Even 128 entries seems excessive, but try to appear
conventional.