Modify handling of /cust/official_op_resize.cfg
The initial handling of /cust/official_op_resize.cfg had been to simply
overwrite the file.  The new handling is to create a replacement file.
The theory is to allow resizing back to the old size should the user
desire to do so.  I don't expect this to be commonly used, but just in
case someone desires to do such a thing.