Implement modified-order GPT restoration mode
The goal of this new mode is to place userdata (/data) as near the start
of flash as possible and slices likely to be merged closer.  The benefit
is space allocated to OP and cust are easier to merge into userdata.
With userdata earlier resizing is possible.  Potentially some of cache
and system could be repurposed as userdata.