Loosen GPT comparison further
The UEFI specification *requires* reserving space for 128 slice entries,
whether or not that many are allocated, and even if the entry count says
fewer are allocated.  As such the entry count can be increased (which may
be required for the proposed /vendor) without invalidating the
comparison.

The comment about the interpretation of the Alternative LBA field was
incorrect.  Update the comment to match what the specification actually
says.