Initial version of split_system
This is a utility for LineageOS, originally aimed at the LGE V20.  There
is a desire to split off a chunk of the system area for reuse as a vendor
area.  Programs in system are supposed to have limited privileges,
programs in vendor are the ones which require full privileges.