Tiny code tweak, amuse clang
Upon looking at this one little bit, puts() is a better choice for this
spot.  There is no need to scan the string for formatting escapes.
Still, clang's warning was poor.  Clang is smart enough to avoid warning
about the more interesting case in report_kdzfile(), yet dumb enough to
warn about this simpler case.

As this was committed in 2019, portions are now Copyright 2019.