Switch to splitting header in single-pass
With most of the header fields identified, switch to decoding all of them
in a single pass.  Having to call struct multiple times is inefficient.
Grabbing everything at once is better.

Still need clues as to the remaining 8 bytes.  Next test is to see
how much of raw_resources is accounted for and how much remains to
explore.