Detect pixels below image bottom border
No instances of this have ever been observed in the wild, but this case
is in theory possible.  The likely interpretation would be to ignore the
extra pixels, but there is also potential they would crash the
bootloader's decoder.  Still since data like this /might/ be allowable,
send up a flare if such is spotted in the wild.